feat: renamed with react-native-rename
