+++ examples/LaftelVideo/android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.laftelvideo",
+    package = "net.video.fabricexample",
 )
 
 android_resource(
     name = "res",
-    package = "com.laftelvideo",
+    package = "net.video.fabricexample",
     res = "src/main/res",
 )
