+++ examples/LaftelVideo/android/app/src/main/jni/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.13)
 
 # Define the library name here.
-project(laftelvideo_appmodules)
+project(fabricexample_appmodules)
 
 # This file includes all the necessary to let you build your application with the New Architecture.
 include(${REACT_ANDROID_DIR}/cmake-utils/ReactNative-application.cmake)
