+++ examples/LaftelVideo/android/app/src/main/jni/MainComponentsRegistry.h
@@ -13,7 +13,7 @@ class MainComponentsRegistry
  public:
   // Adapt it to the package you used for your Java class.
   constexpr static auto kJavaDescriptor =
-      "Lcom/laftelvideo/newarchitecture/components/MainComponentsRegistry;";
+      "Lnet/video/fabricexample/newarchitecture/components/MainComponentsRegistry;";
 
   static void registerNatives();
