+++ examples/LaftelVideo/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">LaftelVideo</string>
+    <string name="app_name">FabricExample</string>
 </resources>
