+++ examples/LaftelVideo/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "LaftelVideo",
-  "displayName": "LaftelVideo"
+  "name": "FabricExample",
+  "displayName": "FabricExample"
 }
