+++ examples/LaftelVideo/ios/FabricExample/Images.xcassets/AppIcon.appiconset/Contents.json
