+++ examples/LaftelVideo/ios/FabricExample/Images.xcassets/Contents.json
