+++ examples/LaftelVideo/ios/FabricExampleTests/Info.plist
