+++ examples/LaftelVideo/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "laftelvideo",
+  "name": "fabricexample",
   "version": "0.0.1",
   "private": true,
   "scripts": {
